+++ laba3/Land/MyLand.cs
@@ -21,7 +21,7 @@ namespace laba3.Land
             TimeForAfterFirstChill = timeForAfterFirstChill;
         }
 
-        public override int RestDuration(int count)
+        public override double RestDuration(int count)
         {
             if (count == 1)
                 return TimeForFirstChill;
